move typings/models around
